+++ src/metworkpy/network/__init__.py
@@ -28,6 +28,7 @@ from .cluster import (
     get_network_group_linkage,
     get_distance_matrix,
 )
+from .centrality import closeness_centrality_subset
 
 __all__ = [
     "create_metabolic_network",
@@ -52,4 +53,5 @@ __all__ = [
     "get_network_group_clustering",
     "get_network_group_linkage",
     "get_distance_matrix",
+    "closeness_centrality_subset",
 ]